+++ config/custom-command.php
@@ -1,0 +1,17 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Action & Data Transfer Object Settings
+    |--------------------------------------------------------------------------
+    |
+    | This is the namespace all you Eloquent Actions will reside
+    | 
+    */
+    'namespace' => [
+        'action' => 'App\\Actions\\',
+        'dto' => 'App\\DTO\\',
+    ]
+
+];