+++ tests/helpers.php
@@ -1,0 +1,34 @@
+<?php
+
+/**
+ * Get / set the specified configuration value.
+ *
+ * If an array is passed as the key, we will assume you want to set an array of values.
+ *
+ * @param  array|string|null  $key
+ * @param  mixed  $default
+ * @return mixed
+ */
+function config($key = null, $default = null)
+{
+    if (is_null($key)) {
+        return app('config');
+    }
+
+    if (is_array($key)) {
+        return app('config')->set($key);
+    }
+
+    return app('config')->get($key, $default);
+}
+
+/**
+ * Get the configuration path.
+ *
+ * @param  string $path
+ * @return string
+ */
+function config_path($path = '')
+{
+    return app()->basePath() . '/config' . ($path ? '/' . $path : $path);
+}