+++ CHANGELOG.md
@@ -5,4 +5,7 @@ All notable changes to Laravel Command Generator will be documented in this file
 ## Version 1.0.0 (2021-12-14)
 
 ### Added
-- Initial commit.
+- add 2 command
+    - action class
+    - data transfer object class
+- custom class namespace